requirements.txt
# --- Core data + IO ---
numpy==2.3.3
pandas==2.3.2
pyarrow==22.0.0
requests==2.32.5
python-dotenv==1.2.1
pydantic==2.12.2

# --- Database ---
SQLAlchemy==2.0.43
psycopg[binary]==3.2.10

# --- Modeling / stats (panel regression + FE) ---
scikit-learn==1.7.2
statsmodels==0.14.5
scipy==1.16.3
linearmodels==6.1

# --- Visualization (static + exportable) ---
matplotlib==3.10.6
seaborn==0.13.2
plotly==6.5.1
kaleido==1.2.0

# --- Notebook orchestration + export ---
jupyterlab==4.5.1
ipykernel==7.1.0
papermill==2.6.0
nbconvert==7.16.6
playwright==1.56.0

# --- PDF assembly option (non-notebook PDFs, if you choose later) ---
reportlab==4.4.7
Pillow==12.1.0
Jinja2==3.1.6

ruff==0.13.2
black==25.9.0
pytest==9.0.2
mypy==1.18.2
pre-commit==4.5.1
